+++ pkg/resourcemanager/config/env.go
@@ -56,19 +56,19 @@ func ParseEnvironment() error {
 	authorizationServerURL = azureEnv.ActiveDirectoryEndpoint
 	baseURI = azureEnv.ResourceManagerEndpoint // BaseURI()
 
-	locationDefault = envy.Get("AZURE_LOCATION_DEFAULT", "westus2")          // DefaultLocation()
-	useDeviceFlow = ParseBoolFromEnvironment("AZURE_USE_DEVICEFLOW")         // UseDeviceFlow()
-	creds.useManagedIdentity = ParseBoolFromEnvironment("AZURE_USE_MI")      // UseManagedIdentity()
-	keepResources = ParseBoolFromEnvironment("AZURE_SAMPLES_KEEP_RESOURCES") // KeepResources()
-	creds.operatorKeyvault = envy.Get("AZURE_OPERATOR_KEYVAULT", "")         // operatorKeyvault()
+	locationDefault = envy.Get("AZURE_LOCATION_DEFAULT", "westus2")                 // DefaultLocation()
+	useDeviceFlow = ParseBoolFromEnvironment("AZURE_USE_DEVICEFLOW", false)         // UseDeviceFlow()
+	creds.useManagedIdentity = ParseBoolFromEnvironment("AZURE_USE_MI", false)      // UseManagedIdentity()
+	keepResources = ParseBoolFromEnvironment("AZURE_SAMPLES_KEEP_RESOURCES", false) // KeepResources()
+	creds.operatorKeyvault = envy.Get("AZURE_OPERATOR_KEYVAULT", "")                // operatorKeyvault()
 	testResourcePrefix = envy.Get("TEST_RESOURCE_PREFIX", "t-"+helpers.RandomString(6))
 	podNamespace, err = envy.MustGet("POD_NAMESPACE")
 	if err != nil {
 		return errors.Wrapf(err, "couldn't get POD_NAMESPACE env variable")
 	}
 	targetNamespaces = ParseStringListFromEnvironment("AZURE_TARGET_NAMESPACES")
-	purgeDeletedKeyVaultSecrets = ParseBoolFromEnvironment("PURGE_DELETED_KEYVAULT_SECRETS")
-	recoverSoftDeletedKeyVaultSecrets = ParseBoolFromEnvironment("RECOVER_SOFT_DELETED_KEYVAULT_SECRETS")
+	purgeDeletedKeyVaultSecrets = ParseBoolFromEnvironment("PURGE_DELETED_KEYVAULT_SECRETS", false)
+	recoverSoftDeletedKeyVaultSecrets = ParseBoolFromEnvironment("RECOVER_SOFT_DELETED_KEYVAULT_SECRETS", true)
 
 	operatorMode, err = ParseOperatorMode(envy.Get("AZURE_OPERATOR_MODE", OperatorModeBoth.String()))
 	if err != nil {
@@ -137,8 +137,9 @@ func GetExpectedConfigurationVariables() []ConfigRequirementType {
 	return []ConfigRequirementType{RequireClientID, RequireClientSecret, RequireTenantID, RequireSubscriptionID}
 }
 
-func ParseBoolFromEnvironment(variable string) bool {
-	env := envy.Get(variable, "0")
+func ParseBoolFromEnvironment(variable string, defaultValue bool) bool {
+	defaultValueStr := fmt.Sprintf("%t", defaultValue)
+	env := envy.Get(variable, defaultValueStr)
 	value, err := strconv.ParseBool(env)
 	if err != nil {
 		log.Printf("WARNING: invalid input value specified for %q, expected bool, actual: %q. Disabling\n", variable, env)
