+++ pkg/secrets/keyvault/client_test.go
@@ -9,18 +9,15 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/Azure/azure-sdk-for-go/services/keyvault/mgmt/2018-02-14/keyvault"
-	"github.com/Azure/go-autorest/autorest/to"
-	"github.com/gofrs/uuid"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/pkg/errors"
 
 	"github.com/Azure/azure-service-operator/controllers"
 	"github.com/Azure/azure-service-operator/pkg/resourcemanager/config"
 	kvhelper "github.com/Azure/azure-service-operator/pkg/resourcemanager/keyvaults"
 	"github.com/Azure/azure-service-operator/pkg/secrets"
 	keyvaultsecrets "github.com/Azure/azure-service-operator/pkg/secrets/keyvault"
+	testcommon "github.com/Azure/azure-service-operator/test/common"
 )
 
 func getExpectedSecretName(secretKey secrets.SecretKey, namingScheme secrets.SecretNamingVersion) string {
@@ -60,7 +57,7 @@ var _ = Describe("Keyvault Secrets Client", func() {
 		objID, err = kvhelper.GetObjectID(ctx, config.GlobalCredentials(), config.GlobalCredentials().TenantID(), config.GlobalCredentials().ClientID())
 		Expect(err).NotTo(HaveOccurred())
 
-		err = controllers.CreateVaultWithAccessPolicies(
+		err = testcommon.CreateVaultWithAccessPolicies(
 			ctx,
 			config.GlobalCredentials(),
 			resourceGroupName,
@@ -263,12 +260,12 @@ var _ = Describe("Keyvault Secrets soft delete", func() {
 		var err error
 		// Initialize service principal ID to give access to the keyvault
 		kvManager = kvhelper.NewAzureKeyVaultManager(config.GlobalCredentials(), nil)
-		softDeleteKVName = controllers.GenerateTestResourceNameWithRandom("kvsoftdel", 5)
+		softDeleteKVName = controllers.GenerateTestResourceNameWithRandom("kv", 5)
 
 		objID, err = kvhelper.GetObjectID(ctx, config.GlobalCredentials(), config.GlobalCredentials().TenantID(), config.GlobalCredentials().ClientID())
 		Expect(err).NotTo(HaveOccurred())
 
-		err = createKeyVaultSoftDeleteEnabled(
+		err = testcommon.CreateKeyVaultSoftDeleteEnabled(
 			ctx,
 			config.GlobalCredentials(),
 			resourceGroupName,
@@ -285,7 +282,7 @@ var _ = Describe("Keyvault Secrets soft delete", func() {
 		Expect(err).NotTo(HaveOccurred())
 	})
 
-	FIt("Should delete and recreate soft-delete key when recreate soft-deleted key option is enabled", func() {
+	It("Should delete and recreate soft-delete key when recreate soft-deleted key option is enabled", func() {
 		Context("create, delete, and recreate soft deleted key", func() {
 			secretName := "kvsecret" + strconv.FormatInt(GinkgoRandomSeed(), 10)
 			client := keyvaultsecrets.New(
@@ -323,40 +320,3 @@ var _ = Describe("Keyvault Secrets soft delete", func() {
 		})
 	})
 })
-
-func createKeyVaultSoftDeleteEnabled(ctx context.Context, creds config.Credentials, resourceGroupName string, vaultName string, location string, objectID *string) error {
-	vaultsClient, err := kvhelper.GetKeyVaultClient(creds)
-	if err != nil {
-		return errors.Wrapf(err, "couldn't get vaults client")
-	}
-
-	id, err := uuid.FromString(creds.TenantID())
-	if err != nil {
-		return errors.Wrapf(err, "couldn't convert tenantID to UUID")
-	}
-
-	accessPolicies, err := controllers.CreateKeyVaultTestAccessPolicies(creds, objectID)
-	if err != nil {
-		return nil
-	}
-
-	params := keyvault.VaultCreateOrUpdateParameters{
-		Properties: &keyvault.VaultProperties{
-			TenantID:         &id,
-			AccessPolicies:   &accessPolicies,
-			EnableSoftDelete: to.BoolPtr(true),
-			Sku: &keyvault.Sku{
-				Family: to.StringPtr("A"),
-				Name:   keyvault.Standard,
-			},
-		},
-		Location: to.StringPtr(location),
-	}
-
-	future, err := vaultsClient.CreateOrUpdate(ctx, resourceGroupName, vaultName, params)
-	if err != nil {
-		return err
-	}
-
-	return future.WaitForCompletionRef(ctx, vaultsClient.Client)
-}
